pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "bimodal-empathy"
version = "0.1.0"
description = "Bimodal affective alignment: text (GoEmotions) + face (FER2013) + FLAN-T5"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "transformers>=4.36.0",
    "huggingface-hub>=0.20.0",
    "streamlit>=1.29.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "datasets>=2.16.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "python-pptx>=0.6.23",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
